proj_clustering/pyproject.toml
[tool.poetry]
name = "proj-clustering"
version = "0.1.0"
description = ""
authors = ["Tor Thanawat <60775527+thanawat-hub@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.1"
streamlit = "^1.33.0"
google-cloud-bigquery = "^3.20.1"
pymysql = "^1.1.0"
sqlalchemy = "^2.0.29"
db-dtypes = "^1.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
